增加数据源管理功能； 定义数据源接口； 添加数据库类型的数据源插件； 优化序列化功能； 优化数据库连接工具，实现事务管理功能； 数据表实体类添加对某个字段加解密功能；

+++ BIDesigner/graphicsmanager.cpp
@@ -205,6 +205,9 @@ QString GraphicsManager::removeGraphic(const UserPluginDO &data)
             graphic = item;
         }
     }
+    if (graphic == nullptr) {
+        return "";
+    }
     auto graphicId = graphic->id();
     if (graphic) {
         if (!ConfigMaster::instance()->userPlugin->deleteById(data)){
